adjusted to allow for location based chatting. Need to verify if it works...

+++ server/database/migrations/1647455824491-AddRoomsAndMessages.ts
@@ -18,11 +18,11 @@ export class AddRoomsAndMessages1647455824491 implements MigrationInterface {
           },
           {
             name: 'latitude',
-            type: 'int',
+            type: 'float',
           },
           {
             name: 'longitude',
-            type: 'int',
+            type: 'float',
           },
           {
             name: 'roomkey',
